feat: Split picker in multiple files
