+++ src/drive/web/modules/services/components/Picker/index.js
@@ -1,0 +1,1 @@
+export { default } from './Picker'
